Add support to access multiple rows at once
When saving word indexs (w*.idx) often multiple words of the same length
will need to be accessed. This implements a new method that allows that
in an efficient way.

Note: this removes the INDEX_MARK_DELETED mechanism to mark deleted
entries. Entries are now deleted using empty lines again. This makes the
batch handling much simpler. If a good reason exists that we should keep
it, it can be readded.