fix handling of loading auth backend
When a non existing auth backend was configured, the action router ran
into an infinie loop exception. The reason was that the denied action
required a configured auth system, but denying access should always
work.

Interestingly the problem did not occur when the auth backend signalled
a failure to load. This was because the auth backend was not properly
deinitialized. This is now done.

To aid debugging similar problems, fatal errors are now logged through
the logging mechanism in the action router