avoid timeouts in tests
For some reason it seems that CLI tests on github run with a execution
time limit that can't be disabled by setting the appropriate ini value.
Even more odd, the problem seems to exist on windows only (even though
the same limit is set on both OSes). In any case, this simply resets the
time limit for each test run and that seems to fix the timeout issue.