fix autoloading for tests
This replaces my attempt in 12ebce974be688ac34bf4389645d5fc7baa29cc7

The canonical namespace for DokuWiki core tests is dokuwiki/test/ and
this is mapped to _test/tests now in the autoloader.

This means the majority of tests is in the dokuwiki/test/inc namespace.

Mockfiles are located at _test/mock and have the namespace
dokuwiki/test/mock - if that's good or bad is debatable. I simply kept
it as it always has been. But there might be an argument for having mock
objects closer to the tests that use them (eg. right next to the test
files).