Fix handling of unknown array settings
When a plugin used an array setting and was then installed, the old
configuration can remain in local.php. It is handled by the
SettingUndefined class. However this class used the base out() mechanism
which didn't know about arrays. The result was a fatal error resulting
in a completely empty config file.

This moves the error handling from the SettingArray class to the base
class which fixes the original problem.

It also adds a check if the Writer is about to write a completely empty
config file and refuses to do so, throwing an Exception.

Probably related to #3777