Check Basic Auth scheme in Authorization header
Prior to this, auth_setup() would simply assume that the Authorization
header was using the Basic auth scheme, but there are other available
ones, which could result in incorrect processing of the header's data.

We now specifically check that we have the `Basic` scheme, and only then
perform the base64_decode to get the username and password.