Add JSON based alternative to XMLRPC
XMLRPC is a rather outdated and old-fashioned protocol not much in use
anymore. Developers prefer simpler, JSON based APIs.

This adds a new "JSONRPC" API. Basically it exposes exactly the same
method calls as the XMLRPC API but using JSON instead of XML. It's not a
classical REST API, but should be just as easy to use for developers.

Here is an example call using CURL:

  curl http://localhost/dokuwiki/lib/exe/jsonrpc.phs \
     -H 'Content-Type: application/json' \
     -H "Authorization: Bearer $token" \
     -d '["wiki"]'

Please note that the above uses the token auth implemented in #2432.
Authentication via basic auth or cookies would work as well.