readd language file
This file had been deleted because I didn't know how to properly remove
the link in it. But looking a bit more at the original it seems that the
link was never correct in the first place which explains why I had
trouble with it.

I now completely removed it. Since I don't speak the language, the
remaining sentence might be weird or miss a word. Native speakers like
@qezwan or @Shvan-ali should have a look and maybe submit an update.

Readding this file should fix #3988 updating endlessly.