deprecate ptln()
This method was used to ensure some basic readability in the created
HTML sources long before Firebug and later the builtin inspector tools
in browsers pretty printed the HTML for you. Today, this is no longer
needed.

This adds a custom rector rule to automatically change all occurances to
echo statements.