Refactor autoloading, fix #4048
This refactors the auto loading stuff into an anonymous class and cleans
it up a bit.

It also ensures that plugin classes are not autoloaded when the plugin
is disabled. This however only works after the plugin controller has
been initialized.

We currently reference some classes of the config plugin in out
deprecated.php file resulting in these classes being loaded before the
plugin controller. Not a big deal I guess.