More resilence when talking to plugin repo fixes #4135
When the plugin repository answers but returns non-data (as can happen
when the database isn't available, as we have learned yesterday), the
extension manager still tried to deserialize the data.

This changes all communication from php-serialized to JSON encoded data.
When JSON-decoding fails, the data is ignored and an error message is
shown. Failure data like this will not be cached.