fall back to empty (root) dir for base dir detection
When a basedir could not be detected, the default previously was '.'
resulting in a valid but weird URL (http://example.com/./doku.php). We
now default to an empty dir, resulting in a more sensible URL of
http://example.com/doku.php

This should not matter in real web server setups but will be in effect
while testing.