fix overeager optimization in Table handler. fixes #4186
This reverts an optimization introduced in
bcaec9f47d06126b3e653fea89a86d8b6a6cbef8

The number of elements in $this->tableCalls may change during the loop,
so they need to be recounted on every step. To protect it from being
"optimized" again, the loop was changed into a while loop.

Ultimately it should be checked if this method could be optimized in
another way.