Use environment provided NONCE for inline scripts. #3788
When an outside source wants to set a restrictive CSP, it can use a
nonce to allow inline scripts instead of using 'unsafe-inline'. This
nonce can be passed on via the environment variable NONCE and will be
used in tpl_metaheaders to tag our inline JS initializations.

An update to the cspheaders plugin should be made to provide a nonce as
well.