move JS detection to default script
This moves the little helper that will remove a "no-js" class from the
header element as soon as JavaScript is detected from the dokuwiki
template to our default inline script that also initializes the JSINFO
array.

This ensures that this inline script is run with a nonce (if available).
See #3788 for more infor