Output more version data on the admin screen
Bug reports are often lacking crucial information about the environment
the wiki is run on and novice users often do not know how to obtain this
data.

This patch gathers some environmental information and outputs it next to
the DokuWiki version number on the admin screen.

Info included (if obtainable):

* PHP version
* Linux Distribution name and version
* Operating System and Version (refers to kernel version on linux)
* The PHP SAPI (modphp, cgi, fcgi)
* If running on kubernetes or docker