Fix display of media manager namespace hierarchy
When the media manager is called with ns parameter set to a value higher
than the last namespace defined in the wiki, the non-existing namespace
is not added to the hierarchy.

This behavior is not consistent with what happens when the namespace's
name is lower. In this case, an entry is inserted in the tree at the
appropriate location.

Fixed by appending the temporary namespace at the end of the tree if it
was not inserted by the search loop.

Fixes #4276