Unique identifiers for editor toolbar buttons
This change adds unique ids to the editor toolbar buttons. It is a proposed solution for issue #4335.

In addition, it changes the default template to make use of these identifiers for grouping of the buttons.

(a small increase in the size of those buttons is also added, as they are really rather small!)