Extension Manager: set a larger timeout
When loading large plugins, like the dw2pdf plugin, a timout of 25
seconds may not be enough on slow connections. This increases the limit
to 5 second short of 4 minutes.

On CLI this should not be an issue at all, on web SAPIs we need to
increase the execution time first, only if that succeeds (might be
disabled on some hosts) the limit is increased.

This replaces PR #4377