allow passing current revision to mediaFN
This allows passing the current revision's timestamp as $rev parameter
and get the path to the file in data/meta. Previously you had to always
pass an empty revision to access the current file.

This makes it easier to use the return of MediaChangeLog::getRevisions()
to access any revision.

This will also fix the issues encountered in #4147