check permissions for admin menu item earlier
I am not sure why moving the context visibility check to a later stage
suddenly surfaced the issue of $INFO not being set in the
general_html_test. I would have expected the same issue before the
changes of 33792c0e7da35449c4591cfbb2b77377e2f465b1

In any case it makes not much sense to do the permission checks there
anyway. Doing it in the constructor makes more sense since this is not a
context sensitive check.