fix OS checks. fixes #4446 and #4445
Access to /etc might be restricted. We simply ignore the errors and
pretend the file does not exist.

Strictly speaking, the files are not in ini format. But parsing them as
ini mostly works, except for some comments (on synology), so we strip
the comments first.

This also makes sure the correct file is read on synology.