ExtensionManager: get base from plugin classes. fixes #4484
When installing a plugin that has no plugin.info.txt, we need to figure
out where it belongs from other data. Previously we checked the archive
name or folder within. This will fail for archives generated from github
branches as the plugin would be named "dokuwiki-plugin-foobar" or "master".

This implements looking for known plugin type php files, parsing them
with PHP's parser and extracting the class name. The base name is in
that class name.

Seems to work fine in my unit and manual tests.