SearchIndex: TupleOps now work with frequencies of 1
We have indexes where we simply track that a relation between entity and
token exists, but there is no frequency. The frequency is always 1.
For those indexes we do not store *1 as frequency but omit it
completely.

The TupleOps class now can work with such indexes and will also store
frequencies of 1 like this.